+++ NEWS.md
@@ -1,8 +1,12 @@
 # simIDM 0.0.5.9013
 
-- ExpSurvOS now returns 0 instead of NaN for large values of t (#73, @holgstr)
+### Bug Fixes
 
-# simIDM 0.0.5.9011
+- `ExpSurvOS` now returns 0 instead of NaN for large values of t.
+- `getSimulatedData` now also works when there are no transitions from progression to death, similarly for `getOneClinicalTrial` (which now warns if there are no such transitions at all).
+
+
+# simIDM 0.0.5
 
 - First CRAN version of the package.
 - The package simulates illness-death models with constant, Weibull or piecewise constant transition hazards.